refactor: extract Formbricks tracking into separate hook
- Create useFormbricksOptInTracking hook for delayed survey triggering
- Update useFeatureOptInBanner to call the new hook
- Export helper functions from the new hook for reuse
